+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Features
 
+- Mount fastmcp streamable http server
 - Build local oauth authorization server
 - Add mcp config validation